+++ src/layer/psroipooling.cpp
@@ -60,10 +60,10 @@ int PSROIPooling::forward(const std::vector<Mat>& bottom_blobs, std::vector<Mat>
     // For each ROI R = [x y w h]: avg pool over R
     const float* roi_ptr = roi_blob;
 
-    float roi_x1 = round(roi_ptr[0]) * spatial_scale;
-    float roi_y1 = round(roi_ptr[1]) * spatial_scale;
-    float roi_x2 = round(roi_ptr[2] + 1.f) * spatial_scale;
-    float roi_y2 = round(roi_ptr[3] + 1.f) * spatial_scale;
+    float roi_x1 = static_cast<float>(round(roi_ptr[0]) * spatial_scale);
+    float roi_y1 = static_cast<float>(round(roi_ptr[1]) * spatial_scale);
+    float roi_x2 = static_cast<float>(round(roi_ptr[2] + 1.f) * spatial_scale);
+    float roi_y2 = static_cast<float>(round(roi_ptr[3] + 1.f) * spatial_scale);
 
     float roi_w = std::max(roi_x2 - roi_x1, 0.1f);
     float roi_h = std::max(roi_y2 - roi_y1, 0.1f);
@@ -82,10 +82,10 @@ int PSROIPooling::forward(const std::vector<Mat>& bottom_blobs, std::vector<Mat>
             {
                 const float* ptr = bottom_blob.channel((q * pooled_height + ph) * pooled_width + pw);
 
-                int hstart = floor(roi_y1 + (float)(ph) * bin_size_h);
-                int wstart = floor(roi_x1 + (float)(pw) * bin_size_w);
-                int hend = ceil(roi_y1 + (float)(ph + 1) * bin_size_h);
-                int wend = ceil(roi_x1 + (float)(pw + 1) * bin_size_w);
+                int hstart = static_cast<int>(floor(roi_y1 + (float)(ph) * bin_size_h));
+                int wstart = static_cast<int>(floor(roi_x1 + (float)(pw) * bin_size_w));
+                int hend = static_cast<int>(ceil(roi_y1 + (float)(ph + 1) * bin_size_h));
+                int wend = static_cast<int>(ceil(roi_x1 + (float)(pw + 1) * bin_size_w));
 
                 hstart = std::min(std::max(hstart, 0), h);
                 wstart = std::min(std::max(wstart, 0), w);
