+++ src/layer/quantize.cpp
@@ -35,7 +35,7 @@ int Quantize::load_param(const ParamDict& pd)
 
 static inline signed char float2int8(float v)
 {
-    int int32 = round(v);
+    int int32 = static_cast<int>(round(v));
     if (int32 > 127) return 127;
     if (int32 < -127) return -127;
     return (signed char)int32;
